+++ lib/spout/tests.rb
@@ -6,6 +6,7 @@ require 'json'
 require 'spout/tests/json_validation'
 require 'spout/tests/variable_type_validation'
 require 'spout/tests/variable_name_uniqueness'
+require 'spout/tests/variable_name_match'
 require 'spout/tests/domain_existence_validation'
 require 'spout/tests/domain_format'
 require 'spout/tests/domain_name_uniqueness'
@@ -16,6 +17,7 @@ module Spout
     include Spout::Tests::JsonValidation
     include Spout::Tests::VariableTypeValidation
     include Spout::Tests::VariableNameUniqueness
+    include Spout::Tests::VariableNameMatch
     include Spout::Tests::DomainExistenceValidation
     include Spout::Tests::DomainFormat
     include Spout::Tests::DomainNameUniqueness